Application/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]
snowflake-connector-python==3.5.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0

# Streamlit Frontend Dependencies
streamlit==1.28.1
requests==2.31.0